requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
requests==2.31.0
tavily-python==0.3.1
langchain
langchain-community
langchain-core